LayoutTests/inspector/debugger/script-formatter-console-expected.txt

CONSOLE MESSAGE: line 10: Test message.
Tests that the script formatting changes console line numbers.

Debugger was enabled.

Running: testConsoleMessagesForFormattedScripts
script-formatter-console.html:10
Pre-format row message list:
["9"]
["9"]
script-formatter-console.html:13
Post-format row message list:
["12"]
["12"]
Debugger was disabled.
